// src/app/models/albums - artists.interface.ts
import { Followers } from 'src/app/models/search.interface';

export interface NewReleases {
  albums: Albums;
}

export interface Albums {
  href: string;
  items: Album[];
  limit: number;
  next: string | null;
  offset: number;
  previous: string | null;
  total: number;
}

export interface Album {
  album_type: string;
  artists: Artist[];
  available_markets: string[];
  external_urls: ExternalUrls;
  href: string;
  id: string;
  images: Image[];
  name: string;
  release_date: Date;
  release_date_precision: string;
  total_tracks: number;
  type: string;
  uri: string;
}
export interface Artist {
  images: Image[];
  external_urls: ExternalUrls;
  href: string;
  id: string;
  name: string;
  type: ArtistType;
  uri: string;
  isFollowing: boolean;
  followers: Followers;
  genres?: string[];
  popularity?: number;
}

export interface ExternalUrls {
  spotify: string;
}

export enum ArtistType {
  Artist = 'artist',
}

export interface Image {
  height: number;
  url: string;
  width: number;
}